bmad_tui/requirements.txt
textual>=0.60.0
pyyaml>=6.0
pyfiglet>=1.0.2
